refactor: create KeyHelper which contains all functionality related to TCK keys. Reworked to use key type enum for checking
